+++ src/config-examples/ors-map-filters-example.js
@@ -87,6 +87,12 @@ const filters = [
         value: 'driving-car'
       }
     ],
+    visibleWhen: [
+      {
+        ref: 'profile',
+        value: 'driving-car'
+      }
+    ],
     availableOnModes: [constants.modes.directions],
     useInServices: [constants.services.directions]
   },