+++ src/pages/maps/maps.js
@@ -163,6 +163,11 @@ export default {
       let center = null
       if (this.$store.getters.appRouteData.options.center) {
         center = this.$store.getters.appRouteData.options.center
+      } else { // set the map center based on the only place available (if it is the case)
+        const filledPlaces = this.lodash.filter(this.$store.getters.appRouteData.places, (p) => { if (!p.isEmpty()) { return p } })
+        if (filledPlaces.length === 1) {
+          center = {lat: filledPlaces[0].lat, lng: filledPlaces[0].lng}
+        }
       }
       return center
     },